requirements.txt
streamlit==1.35.0
face_recognition==1.4.0
face_recognition_models==0.3.0
dlib==19.24.0
cmake>=3.26
numpy>=1.24
Pillow==9.5.0
scikit-learn==1.4.2
opencv-python==4.9.0.80
ftfy==6.2.0
regex==2024.4.28
tqdm==4.66.4
torch==2.2.2
torchvision==0.17.2
git+https://github.com/openai/CLIP.git
